// frontend/src/app/api/models/inicio.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DestinoTarjeta } from '../models/destino-tarjeta';
import { GuiaTarjeta } from '../models/guia-tarjeta';
import { ImagenPublica } from '../models/imagen-publica';
import { ItinerarioTarjeta } from '../models/itinerario-tarjeta';
import { TipoAventuraTarjeta } from '../models/tipo-aventura-tarjeta';
export interface Inicio {
  destinos: Array<DestinoTarjeta>;
  guias: Array<GuiaTarjeta>;
  hero: ({
'titular': string;
'subtitulo': string;
'imagen': ImagenPublica;
} | null);
  itinerarios: Array<ItinerarioTarjeta>;
  tipos: Array<TipoAventuraTarjeta>;
}